+++ Library/ISCore/ISTcp.h
@@ -9,7 +9,7 @@ class ISTcp
 public:
 	static bool IsValidQuery(const QByteArray &ByteArray, QVariantMap &VariantMap, QString &ErrorString); //Проверить валидность запроса
 	static bool IsValidAnswer(const QByteArray &ByteArray, QVariantMap &VariantMap, QString &ErrorString); //Проверить валидность ответа на запрос
-	static long GetQuerySizeFromBuffer(QByteArray &ByteArray); //Получить размер запроса из его содержимого
+	static int GetQuerySizeFromBuffer(QByteArray &ByteArray); //Получить размер запроса из его содержимого
 	static QByteArray Crypt(const std::vector<unsigned char> &Key, const QVariantMap &VariantMap); //Шифрование данных
 };
 //-----------------------------------------------------------------------------
